backend/requirements.txt
Flask
python-dotenv
sentence-transformers
pytesseract
flask-cors
numpy<2

# apt-get install tesseract-ocr
